Fix correctness bugs and harden error handling found in code review (#50)
Fixes several issues discovered during a review of the C++ core:

- GetVectorByRowid hardcoded getDataByLabel<float>, so reading the vector
  column back from a bfloat16/float16 table over-read the stored buffer and
  returned garbage. Now dispatches on the stored vector type and dequantizes
  back to f32.
- InitVirtualTable leaked the VirtualTable and left *ppVTab dangling when
  loading the index file failed. Now deletes it and clears the pointer.
- knn_search's ef override mutated shared index state and leaked into
  subsequent queries (and raced under concurrency). ef_ is now snapshotted
  and restored via absl::Cleanup.
- xColumn returned the invalid SQLITE_NOTFOUND code; now returns SQLITE_ERROR.
- IndexOptions::FromString silently ignored malformed tokens; the whole
  option string is now validated as comma-separated key=value pairs.
- KnnParam held a non-owning VectorView that could dangle for temporary blob
  arguments; it now owns a Vector copy.

Also: fix stale doc comment in vector_space.h, remove dead
IndexConstraintUsage enum and a stale FloorLog2 comment, use size_t loop
counters in Normalize_Scalar, and avoid a double ToBlob() call in
VectorFromJson.

Adds regression tests for the quantized vector read-back and malformed
index-option parsing.

+++ vectorlite/ops/ops.cpp
@@ -810,8 +810,7 @@ HWY_AFTER_NAMESPACE();
 namespace vectorlite {
 namespace ops {
 
-// This macro declares a static array used for dynamic dispatch; it resides in
-// the same outer namespace that contains FloorLog2.
+// This macro declares a static array used for dynamic dispatch.
 HWY_EXPORT(InnerProductImplF32);
 HWY_EXPORT(InnerProductImplBF16);
 HWY_EXPORT(InnerProductImplF16);
@@ -919,12 +918,12 @@ HWY_DLLEXPORT float L2DistanceSquared(const float* v1,
 // Not sure whether compiler will do auto-vectorization for this function.
 HWY_DLLEXPORT void Normalize_Scalar(float* HWY_RESTRICT inout, size_t size) {
   float norm = 0.0f;
-  for (int i = 0; i < size; i++) {
+  for (size_t i = 0; i < size; i++) {
     float data = inout[i];
     norm += data * data;
   }
   norm = 1.0f / (sqrtf(norm) + 1e-30f);
-  for (int i = 0; i < size; i++) {
+  for (size_t i = 0; i < size; i++) {
     inout[i] = inout[i] * norm;
   }
   return;
@@ -933,12 +932,12 @@ HWY_DLLEXPORT void Normalize_Scalar(float* HWY_RESTRICT inout, size_t size) {
 HWY_DLLEXPORT void Normalize_Scalar(hwy::bfloat16_t* HWY_RESTRICT inout,
                                     size_t size) {
   float norm = 0.0f;
-  for (int i = 0; i < size; i++) {
+  for (size_t i = 0; i < size; i++) {
     float data = hwy::F32FromBF16(inout[i]);
     norm += data * data;
   }
   norm = 1.0f / (sqrtf(norm) + 1e-30f);
-  for (int i = 0; i < size; i++) {
+  for (size_t i = 0; i < size; i++) {
     inout[i] = hwy::BF16FromF32(hwy::F32FromBF16(inout[i]) * norm);
   }
   return;
@@ -947,12 +946,12 @@ HWY_DLLEXPORT void Normalize_Scalar(hwy::bfloat16_t* HWY_RESTRICT inout,
 HWY_DLLEXPORT void Normalize_Scalar(hwy::float16_t* HWY_RESTRICT inout,
                                     size_t size) {
   float norm = 0.0f;
-  for (int i = 0; i < size; i++) {
+  for (size_t i = 0; i < size; i++) {
     float data = hwy::F32FromF16(inout[i]);
     norm += data * data;
   }
   norm = 1.0f / (sqrtf(norm) + 1e-30f);
-  for (int i = 0; i < size; i++) {
+  for (size_t i = 0; i < size; i++) {
     inout[i] = hwy::F16FromF32(hwy::F32FromF16(inout[i]) * norm);
   }
   return;
